remove extension variable
Issues if within the name of the original files are present different "." (ex: name.red.fcs). In this case, indeed, the generated file take the .red extension and not the .fcs one. To avoid this issue i removed the extension variable (line 43 and 61) and added the string "fcs" just after the "." at line 45 and 63